% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CB_FeatureImp.R
\name{CB_FeatureImp}
\alias{CB_FeatureImp}
\title{Feature importance plot}
\usage{
CB_FeatureImp(
  model,
  train,
  target.name,
  itr = 10,
  task = "regression",
  parallel = F,
  pkg = NULL
)
}
\arguments{
\item{model}{A prediction model (Classification or regression)}

\item{train}{Training dataset (data frame) that is used to building model}

\item{target.name}{Name of target label name in train dataset}

\item{itr}{number of iteration of experiment. default is 10}

\item{task}{Prediction task. "regression" (default) or "classification".}

\item{parallel}{Apply parallel processing. T of F (default)}

\item{pkg}{When parallel=T, supported package name for model should be given.}
}
\value{
[graph object] Feature importance plot
}
\description{
This plot shows feature importance for given dataset.
feature importance is composed of feature power (contribution) and interaction power
}
\examples{
library(CBFI)
# for regression
data("Boston", package = "MASS")
model1 <- lm(medv ~ ., data = Boston)
result1 <- CB_FeatureImp(model1, train=Boston, target.name="medv", itr=50, task="regression")
print(result1)

# for classification
library(e1071)
model2 <- svm(Species~., data=iris)
result2 <- CB_FeatureImp(model2, train=iris, target.name="Species", itr=50, task="classification")
print(result2)

# for classification (with parallel processing)
library(e1071)
model3 <- svm(Species~., data=iris)
result3 <- CB_FeatureImp(model3, train=iris, target.name="Species", itr=50, task="classification", parallel=T, pkg="e1071")
print(result3)

}
