% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CB_FItable.R
\name{CB_FItable}
\alias{CB_FItable}
\title{Create a feature interaction table}
\usage{
CB_FItable(
  model,
  train,
  target.name,
  itr = 50,
  task = "regression",
  class = "_all_",
  parallel = F,
  pkg = NULL
)
}
\arguments{
\item{model}{A prediction model (Classification or regression)}

\item{train}{Training dataset (data frame) that is used to building model}

\item{target.name}{Name of target label name in train dataset}

\item{itr}{Number of experiments. Default is 50}

\item{task}{Prediction task. "regression" (default) or "classification".}

\item{class}{class name (for classification model)}

\item{parallel}{Apply parallel processing. T of F (default)}

\item{pkg}{When parallel=T, supported package name for model should be given.}
}
\value{
[list] feature interaction & feature importance table
}
\description{
This function create a feature interaction table.
}
\examples{
library(CBFI)
# for regression
data("Boston", package = "MASS")
model1 <- lm(medv ~ ., data = Boston)
result1 <- CB_FItable(model1, Boston, "medv", itr=50, task="regression")
print(result1)

# for classification
library(e1071)
model2 <- svm(Species~., data=iris)
result2 <- CB_FItable(model2, iris, "Species", itr=50, task="classification")
print(result2)

# for classification (parallel processing)
library(e1071)
model3 <- svm(Species~., data=iris)
result3 <- CB_FItable(model3, iris, "Species", itr=50, task="classification",
                      parallel=T, pkg="e1071")
print(result3)

}
