% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CB_plot.FIgraph.R
\name{CB_plot.FIgraph}
\alias{CB_plot.FIgraph}
\title{Draw feature interaction graph}
\usage{
CB_plot.FIgraph(
  gtype = "S",
  FIobj,
  task = "regression",
  class.name = "_all_",
  show.edge.weight = TRUE,
  seed = 100,
  th = NULL
)
}
\arguments{
\item{gtype}{Graph style. 'S' (default) is for static graph and 'I' is for interactive graph.}

\item{FIobj}{Result object from FItable function.}

\item{task}{Type of prediction task, 'regression' (default), 'Classification'}

\item{class.name}{It is for classification task. You can draw feature interaction graph for specific class.}

\item{show.edge.weight}{Decide if edge weight value is displayed or not.}

\item{seed}{According to seed value, The shape of graph is changed.}

\item{th}{Threshold for showing edge. If th is NULL, th is mean of interaction values of all pair of teatures}
}
\value{
Feature interaction graph
}
\description{
This function Draw feature interaction graph.
}
\examples{
library(CBFI)
# for regression
data("Boston", package = "MASS")
model1 <- lm(medv ~ ., data = Boston)
FIobj1 <-CB_FItable(model1, Boston, "medv", itr=50, task="regression")

CB_plot.FIgraph(gtype="S", FIobj1, task="regression", class="_all_",
               show.edge.weight=TRUE, seed=104, th=NULL)
CB_plot.FIgraph(gtype="I", FIobj1, task="regression", class="_all_",
                   show.edge.weight=TRUE, seed=104, th=NULL)#' # for classification

library(e1071)
model2 <- svm(Species~., data=iris)
FIobj2 <-  CB_FItable(model2, iris, "Species", itr=50, task="classification")

# static plot
CB_plot.FIgraph( gtype="S", FIobj=FIobj, task="classification", class="_all_",
                show.edge.weight=TRUE, seed=104)
CB_plot.FIgraph( gtype="S", FIobj=FIobj, task="classification", class="setosa",
                show.edge.weight=TRUE, seed=104)
# interactive plot
CB_plot.FIgraph( gtype="I", FIobj=FIobj, task="classification", class="_all_",
                show.edge.weight=TRUE, seed=104)

# Please ignore warning error after drawing interactive graph.

}
