% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CB_plot.imp.R
\name{CB_plot.imp}
\alias{CB_plot.imp}
\title{Bar plot for feature importance}
\usage{
CB_plot.imp(result, class = "_all_", combine = F)
}
\arguments{
\item{result}{An object from CB_FeatureImp function}

\item{class}{A class name, (for classification model)}

\item{combine}{if combine=F, graph bar split into contribution and interaction (for  regression model)}
}
\value{
[graph object] feature interaction plot
}
\description{
This plot shows amount of interaction between two features
}
\examples{
library(CBFI)
# for regression
data("Boston", package = "MASS")
model1 <- lm(medv ~ ., data = Boston)
result1 <- CB_FeatureImp(model1, train=Boston, target.name="medv", itr=50, task="regression")

CB_plot.imp(result1, combine=F)
CB_plot.imp(result1, combine=T)

# for classification
library(e1071)
model2 <- svm(Species~., data=iris)
result2 <- CB_FeatureImp(model.svm, train=iris, target.name="Species", itr=50, task="classification")
print(result2)
CB_plot.imp(result2, class="_all_")
CB_plot.imp(result2, class="versicolor")

}
