% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CB_plot.interact.R
\name{CB_plot.interact}
\alias{CB_plot.interact}
\title{Bar plot for interaction between two features}
\usage{
CB_plot.interact(result, class = "_all_")
}
\arguments{
\item{result}{An object from CB_FeatureInteract function}

\item{class}{A class name, if result is from classification model}
}
\value{
[graph object] feature interaction plot
}
\description{
This plot shows amount of interaction between two features
}
\examples{
library(CBFI)
# for regression
data("Boston", package = "MASS")
model1 <- lm(medv ~ ., data = Boston)
result1 <- CB_FeatureInteract(model.lm, train=Boston, target.name="medv", F1="lstat", F2=NULL, itr=50, task="regression")

print(result1)
CB_plot.interact(result1)

# for classification
library(e1071)
model2 <- svm(Species~., data=iris)
result2 <- CB_FeatureInteract(model2, train=iris, target.name="Species", F1="Petal.Width", F2=NULL, itr=50, task="classification")
CB_plot.interact(result2, class="_all_")
CB_plot.interact(result2, class="setosa")

result3 <- CB_FeatureInteract(model2, train=iris, target.name="Species", F1="Petal.Width", F2="Petal.Length", itr=50, task="classification")
CB_plot.interact(result3, class="_all_")
CB_plot.interact(result3, class="versicolor")
#'
}
