% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CB_plot.PA.R
\name{CB_plot.PA}
\alias{CB_plot.PA}
\title{Prediction analysis plot for two features}
\usage{
CB_plot.PA(
  model,
  train,
  target.name,
  F1,
  F2,
  itr = 10,
  task = "regression",
  class = "_all_"
)
}
\arguments{
\item{model}{A prediction model (Classification or regression)}

\item{train}{Training dataset (data frame) that is used to building model}

\item{target.name}{Name of target label name in train dataset}

\item{F1}{Name of first feature}

\item{F2}{Name of second feature}

\item{itr}{number of iteration of experiment. default is 10}

\item{task}{Prediction task. "regression" (default) or "classification".}

\item{class}{When task is classification, PA plot can be generated for specific class.
If class is "_all_", PA plot is generated for all class.}
}
\value{
[graph object] Prediction analysis plot
}
\description{
This plot shows both data distribution and contribution for two features.
If prediction task is 'classification', this plot can be generated for specific class.
}
\examples{
library(CBFI)
# for regression
data("Boston", package = "MASS")
model1 <- lm(medv ~ ., data = Boston)
CB_plot.PA(model1, Boston, "medv", F1="dis", F2="rad", itr=10, task="regression")
CB_plot.PA(model1, Boston, "medv", F1="lstat", F2="rm", itr=10, task="regression")

# for classification
library(e1071)
model2 <- svm(Species~., data=iris)
CB_plot.PA(model2, iris, "Species", F1="Petal.Width", F2="Petal.Length", itr=10, task="classification")
CB_plot.PA(model2, iris, "Species", F1="Petal.Width", F2="Petal.Length", itr=10, task="classification", class="setosa")

}
